requirements.txt
streamlit>=1.40
requests>=2.32
pandas>=2.2
pdfplumber>=0.11
PyMuPDF>=1.27
python-docx>=1.1
openpyxl>=3.1
pydantic>=2.0
pyyaml>=6.0
pypdf>=4.0
numpy>=1.26
scikit-learn>=1.4
